rename authenticate API router
