+++ routes/authenticateAPI.js
